moved process* functions to processors.rs
